Rename config to runtimeConfig
