+++ lib/runtime-config.js
@@ -1,9 +1,9 @@
-let env
+let runtimeConfig
 
 export default () => {
-  return env
+  return runtimeConfig
 }
 
 export function setConfig (configValue) {
-  env = configValue
+  runtimeConfig = configValue
 }
